# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is docs of Celestia indexer API.",
        "title": "Swagger Celestia Indexer API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1",
    "basePath": "/v1",
    "paths": {
...
        "/v1/tx": {
            "get": {
                "description": "List transactions info",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "List transactions info",
                "operationId": "list-transactions",
                "parameters": [
                    {
                        "maximum": 100,
                        "type": "integer",
                        "description": "Count of requested entities",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Sort order",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "success",
                            "failed"
                        ],
                        "type": "string",
                        "description": "Comma-separated status list",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "WithdrawValidatorCommission",
                            "WithdrawDelegatorReward",
                            "EditValidator",
                            "BeginRedelegate",
                            "CreateValidator",
                            "Delegate",
                            "Undelegate",
                            "Unjail",
                            "Send",
                            "CreateVestingAccount",
                            "CreatePeriodicVestingAccount",
                            "PayForBlobs"
                        ],
                        "type": "string",
                        "description": "Comma-separated message types list",
                        "name": "msg_type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Time from in unix timestamp",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Time to in unix timestamp",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Block number",
                        "name": "height",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.Tx"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/v1/tx/{hash}": {
            "get": {
                "description": "Get transaction by hash",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get transaction by hash",
                "operationId": "get-transaction",
                "parameters": [
                    {
                        "maxLength": 64,
                        "minLength": 64,
                        "type": "string",
                        "description": "Transaction hash in hexadecimal",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Tx"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/v1/tx/{hash}/events": {
            "get": {
                "description": "Get transaction events",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get transaction events",
                "operationId": "get-transaction-events",
                "parameters": [
                    {
                        "maxLength": 64,
                        "minLength": 64,
                        "type": "string",
                        "description": "Transaction hash in hexadecimal",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.Event"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/v1/tx/{hash}/messages": {
            "get": {
                "description": "Get transaction messages",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get transaction messages",
                "operationId": "get-transaction-messages",
                "parameters": [
                    {
                        "maxLength": 64,
                        "minLength": 64,
                        "type": "string",
                        "description": "Transaction hash in hexadecimal",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.Message"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handler.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "responses.Event": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "height": {
                    "type": "integer",
                    "format": "int64",
                    "example": 100
                },
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 321
                },
                "position": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "time": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-07-04T03:10:57+00:00"
                },
                "tx_id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 11
                },
                "type": {
                    "type": "string",
                    "format": "string",
                    "enum": [
                        "coin_received",
                        "coinbase",
                        "coin_spent",
                        "burn",
                        "mint",
                        "message",
                        "proposer_reward",
                        "rewards",
                        "commission",
                        "liveness",
                        "attestation_request",
                        "transfer",
                        "pay_for_blobs",
                        "redelegate",
                        "withdraw_rewards",
                        "withdraw_commission",
                        "create_validator",
                        "delegate",
                        "edit_validator",
                        "unbond",
                        "tx",
                        "unknown"
                    ],
                    "example": "commission"
                }
            }
        },
        "responses.Message": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "height": {
                    "type": "integer",
                    "format": "int64",
                    "example": 100
                },
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 321
                },
                "position": {
                    "type": "integer",
                    "format": "int64",
                    "example": 2
                },
                "time": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-07-04T03:10:57+00:00"
                },
                "tx_id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 11
                },
                "type": {
                    "type": "string",
                    "format": "string",
                    "enum": [
                        "WithdrawValidatorCommission",
                        "WithdrawDelegatorReward",
                        "EditValidator",
                        "BeginRedelegate",
                        "CreateValidator",
                        "Delegate",
                        "Undelegate",
                        "Unjail",
                        "Send",
                        "CreateVestingAccount",
                        "CreatePeriodicVestingAccount",
                        "PayForBlobs"
                    ],
                    "example": "CreatePeriodicVestingAccount"
                }
            }
        },
...
        "responses.Tx": {
            "type": "object",
            "properties": {
                "codespace": {
                    "type": "string",
                    "format": "string",
                    "example": "sdk"
                },
                "error": {
                    "type": "string",
                    "format": "string",
                    "example": ""
                },
                "events_count": {
                    "type": "integer",
                    "format": "int64",
                    "example": 2
                },
                "fee": {
                    "type": "string",
                    "format": "int64",
                    "example": "9348"
                },
                "gas_used": {
                    "type": "integer",
                    "format": "int64",
                    "example": 4253
                },
                "gas_wanted": {
                    "type": "integer",
                    "format": "int64",
                    "example": 9348
                },
                "hash": {
                    "type": "string",
                    "format": "binary",
                    "example": "652452A670018D629CC116E510BA88C1CABE061336661B1F3D206D248BD558AF"
                },
                "height": {
                    "type": "integer",
                    "format": "int64",
                    "example": 100
                },
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 321
                },
                "memo": {
                    "type": "string",
                    "format": "string",
                    "example": "Transfer to private account"
                },
                "message_types": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "MsgSend",
                        "MsgUnjail"
                    ]
                },
                "messages_count": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "position": {
                    "type": "integer",
                    "format": "int64",
                    "example": 11
                },
                "status": {
                    "type": "string",
                    "format": "string",
                    "enum": [
                        "success",
                        "failed"
                    ],
                    "example": "success"
                },
                "time": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-07-04T03:10:57+00:00"
                },
                "timeout_height": {
                    "type": "integer",
                    "format": "int64",
                    "example": 0
                }
            }
        }
    }
}